pyproject.toml
[project]
name = "aiprojecttemplate"
version = "0.1.0"
description = ""
authors = [
    {name = "ddm9490",email = "ddm9490@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "apache-airflow (>=3.0.1,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = true

packages = [
    { include = "my_template", from = "src" }
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-rich = "^0.2.0"
pre-commit = "^4.2.0"
types-pyyaml = "^6.0.12.20250516"
mypy = "^1.15.0"
jupyter = "^1.1.1"
lxml = "^5.4.0"

[tool.pytest.ini_options]
addopts = "--rich --cov=. --cov-report=html:reports/coverage"
testpaths = ["tests"]
pythonpath = "."

[tool.ruff]
line-length = 100
lint.ignore = ["F401"]

[tool.mypy]
files = ["src/", "tests/"]
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
